server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
pymupdf>=1.25.0
google-generativeai>=0.8.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
